backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
pydantic>=2.0.0
python-multipart==0.0.6
Pillow==10.1.0
python-dotenv==1.0.0
pymysql==1.1.0
supabase==2.3.0
# RAG (optional - mock answers work without these)
sentence-transformers
faiss-cpu
numpy

# ML Models (optional - for image classification)
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.20.0
openai>=1.0.0
google-genai>=0.1.0
anyio<4.0.0
pymongo==4.6.1
